🐛 : Links fixed and stories reducers updated and separated
