+++ client/components/AddPageToStory.js
@@ -35,7 +35,6 @@ const useStyles = makeStyles(theme => ({
 const SavePageToStory = props => {
   console.log('check props for savePageToStory ===', props.stories)
   const stories = props.stories
-  console.log('is this an array', Array.isArray(stories))
 
   const classes = useStyles()
   const [storyId, setStoryId] = React.useState('')
@@ -131,19 +130,19 @@ const SavePageToStory = props => {
                 <em>None</em>
               </MenuItem>
               {/* {console.log('is this story?', Array.isArray(stories))
-                        
+
             }
             {/* { (stories) ? stories.map((story,id)=>{
               console.log('story@@@@@', story)
-            }): null} */}{' '}
-              */}
-              {/* {props.stories.map((story, id) => {
+            }): null} */}
+
+              {props.stories.map((story, id) => {
                 return (
                   <MenuItem key={id} value={story.id}>
                     {story.title}
                   </MenuItem>
                 )
-              })} */}
+              })}
             </Select>
           </FormControl>
         </DialogContent>
