+++ client/components/HomeView.js
@@ -112,7 +112,6 @@ class HomeView extends React.Component {
 
   render() {
     const {classes} = this.props
-    console.log('page!!!', this.props)
 
     return (
       <div className={classes.root}>
@@ -151,7 +150,6 @@ class HomeView extends React.Component {
 }
 
 const mapState = state => {
-  console.log('mapping', state)
   return {
     stickers: state.sticker.stickers,
     user: state.user,
