+++ client/components/PageCard.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import {connect} from 'react-redux'
 import Button from '@material-ui/core/Button'
 import Card from '@material-ui/core/Card'
 import CardActions from '@material-ui/core/CardActions'
@@ -6,13 +7,8 @@ import CardContent from '@material-ui/core/CardContent'
 import CardMedia from '@material-ui/core/CardMedia'
 import CssBaseline from '@material-ui/core/CssBaseline'
 import Grid from '@material-ui/core/Grid'
-import Toolbar from '@material-ui/core/Toolbar'
-import Typography from '@material-ui/core/Typography'
 import {makeStyles} from '@material-ui/core/styles'
 import Container from '@material-ui/core/Container'
-// import Link from '@material-ui/core/Link'
-import {render} from 'enzyme'
-import Canvas from './Canvas'
 import {Link} from 'react-router-dom'
 
 const useStyles = makeStyles(theme => ({
@@ -49,7 +45,6 @@ const useStyles = makeStyles(theme => ({
 
 const PageCard = props => {
   const classes = useStyles()
-  console.log('find user id', props)
   const pages = props.pages
   return (
     <React.Fragment>
@@ -71,7 +66,6 @@ const PageCard = props => {
                     <Button size="small" color="primary">
                       View
                     </Button>
-
                     <Link
                       to={`/${props.userId}/canvas/story/${props.storyId}/page/${page.id}`}
                     >
@@ -89,10 +83,5 @@ const PageCard = props => {
     </React.Fragment>
   )
 }
-
-// const renderCanvas = (id)=> {
-//   console.log('rendering canvas...',id)
-
-// }
 
 export default PageCard
