+++ client/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import {Provider} from 'react-redux'
-import {Router} from 'react-router-dom'
+import {BrowserRouter} from 'react-router-dom'
 import history from './history'
 import store from './store'
 import App from './app'
@@ -11,9 +11,9 @@ import './socket'
 
 ReactDOM.render(
   <Provider store={store}>
-    <Router history={history}>
+    <BrowserRouter history={history}>
       <App />
-    </Router>
+    </BrowserRouter>
   </Provider>,
   document.getElementById('app')
 )
