+++ client/store/index.js
@@ -6,12 +6,14 @@ import user from './user'
 import sticker from './sticker'
 import stories from './stories'
 import allPages from './pages'
+import singleStory from './singleStory'
 import singlePageReducer from './singlePage'
 
 const reducer = combineReducers({
   user,
   sticker,
   stories,
+  singleStory,
   allPages,
   singlePageReducer
 })
@@ -22,3 +24,6 @@ const store = createStore(reducer, middleware)
 
 export default store
 export * from './user'
+export * from './stories'
+export * from './pages'
+export * from './sticker'
