+++ client/store/stories.js
@@ -2,28 +2,13 @@ import axios from 'axios'
 import history from '../history'
 
 //action types
-const GET_ONE_STORY = 'GET_ONE_STORY'
 const GET_ALL_STORIES = 'GET_ALL_STORIES'
-const ADD_STORY = 'ADD_STORY'
 
 //action creator
-const getOneStory = story => {
-  return {
-    type: GET_ONE_STORY,
-    story
-  }
-}
-
 const getAllStories = stories => {
   return {
     type: GET_ALL_STORIES,
     stories
-  }
-}
-const addNewStory = newStory => {
-  return {
-    type: ADD_STORY,
-    newStory
   }
 }
 
@@ -41,75 +26,11 @@ export const fetchAllStories = userId => {
   }
 }
 
-//get one story by id
-export const fetchOneStory = storyId => {
-  return async dispatch => {
-    const res = await axios.get(`/api/stories/${storyId}`)
-    dispatch(getOneStory(res.data))
-  }
-}
-
-//add Story to User Thunk Creator
-export const addStoryToUser = (userId, brandNewStory, fileToUpload) => {
-  return async dispatch => {
-    const data = new FormData()
-    // If file selected
-    if (fileToUpload) {
-      data.append('canvasImage', fileToUpload, fileToUpload.name)
-      axios
-        .post('/api/profile/canvas-img-upload', data, {
-          headers: {
-            accept: 'application/json',
-            'Accept-Language': 'en-US,en;q=0.8',
-            'Content-Type': `multipart/form-data; boundary=${data._boundary}`
-          }
-        })
-        .then(response => {
-          if (200 === response.status) {
-            // If file size is larger than expected.
-            if (response.data.error) {
-              if ('LIMIT_FILE_SIZE' === response.data.error.code) {
-                console.log('FILE TOO BIG')
-              } else {
-                console.log(response.data)
-                // If not the given file type
-              }
-            } else {
-              // Success
-              let fileName = response.data
-              let imageObj = Object.assign(brandNewStory, {
-                coverImage: fileName.location
-              })
-              axios.post(`/api/stories/${userId}/stories`, imageObj)
-              dispatch(addNewStory(response.data))
-              console.log('fileData', fileName)
-              console.log('FILE SUCCESSFULLY LOADED')
-            }
-          }
-        })
-        .catch(error => {
-          // If another error
-          console.error(error)
-        })
-    } else {
-      // if file not selected throw error
-      console.log('UPLOADD SOMETHING')
-    }
-  }
-}
-
 //reducer
-
 export default function(state = initialState, action) {
-  console.log('state$$', state)
-  console.log('action', action)
   switch (action.type) {
     case GET_ALL_STORIES:
       return {...state, allStories: action.stories}
-    case ADD_STORY:
-      return {...state, allStories: [...state.allStories, action.newStory]}
-    case GET_ONE_STORY:
-      return {...state, allStories: action.story}
     default:
       return state
   }
